# client_fix_server.js
const express = require('express');
const cors = require('cors');
const jwt = require('jsonwebtoken');

const app = express();

// JWT Secret
const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key-for-development-only';

// Middleware
app.use(cors({ origin: true, credentials: true }));
app.use(express.json());

// PERSISTENT USER DATA STORAGE - This fixes the "Jenny Wilson" revert issue
let persistentUsers = {
  1: {
    id: 1,
    first_name: "John",
    last_name: "Doe", 
    full_name: "John Doe",
    email: "john.doe@email.com",
    phone: "(+1) 555-0123",
    role: "Developer",
    profile_photo: "https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=150"
  },
  2: {
    id: 2,
    first_name: "Sarah",
    last_name: "Johnson",
    full_name: "Sarah Johnson", 
    email: "sarah.johnson@email.com",
    phone: "(+1) 555-0456",
    role: "Designer",
    profile_photo: "https://images.unsplash.com/photo-1494790108755-2616b612c937?w=150"
  },
  3: {
    id: 3,
    first_name: "Mike",
    last_name: "Chen",
    full_name: "Mike Chen",
    email: "mike.chen@email.com", 
    phone: "(+1) 555-0789",
    role: "Manager",
    profile_photo: "https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150"
  }
};

// PERSISTENT TIMER STORAGE - This fixes the auto-stop issue
let persistentTimers = {};
let dailyLimits = {};

// Load data from files if they exist
function loadPersistentData() {
  try {
    const fs = require('fs');
    const path = require('path');
    
    // Load users
    const usersFile = path.join(__dirname, 'persistent_users.json');
    if (fs.existsSync(usersFile)) {
      const userData = JSON.parse(fs.readFileSync(usersFile, 'utf8'));
      persistentUsers = { ...persistentUsers, ...userData };
      console.log('📂 Loaded user data from persistent storage');
    }
    
    // Load timers
    const timersFile = path.join(__dirname, 'persistent_timers.json');
    if (fs.existsSync(timersFile)) {
      const timerData = JSON.parse(fs.readFileSync(timersFile, 'utf8'));
      persistentTimers = timerData.timers || {};
      dailyLimits = timerData.limits || {};
      console.log('📂 Loaded timer data from persistent storage');
    }
  } catch (error) {
    console.log('⚠️ Using default data (first run)');
  }
}

// Save data to files
function savePersistentData() {
  try {
    const fs = require('fs');
    const path = require('path');
    
    // Save users
    const usersFile = path.join(__dirname, 'persistent_users.json');
    fs.writeFileSync(usersFile, JSON.stringify(persistentUsers, null, 2));
    
    // Save timers
    const timersFile = path.join(__dirname, 'persistent_timers.json');
    fs.writeFileSync(timersFile, JSON.stringify({
      timers: persistentTimers,
      limits: dailyLimits,
      lastSaved: new Date().toISOString()
    }, null, 2));
    
    console.log('💾 Data saved to persistent storage');
  } catch (error) {
    console.log('⚠️ Could not save to persistent storage');
  }
}

// Initialize data
loadPersistentData();

// Authentication middleware
function authenticateToken(req, res, next) {
  const authHeader = req.headers['authorization'];
  const token = authHeader && authHeader.split(' ')[1];

  if (!token) {
    return res.status(401).json({ success: false, message: 'Access token required' });
  }

  try {
    const decoded = jwt.verify(token, JWT_SECRET);
    req.user = decoded;
    console.log(`🔐 Authenticated user: ${decoded.userId} (${decoded.email})`);
    next();
  } catch (error) {
    return res.status(403).json({ success: false, message: 'Invalid token' });
  }
}

// ===== FIX 1: LOGIN API - Consistent with profile =====
app.post('/api/auth/login', (req, res) => {
  const { email, password } = req.body;
  
  console.log(`🔑 Login attempt for: ${email}`);
  
  // Find user by email
  const user = Object.values(persistentUsers).find(u => u.email === email);
  
  if (!user) {
    return res.status(401).json({
      success: false,
      message: 'Invalid credentials'
    });
  }
  
  // Generate token
  const token = jwt.sign({ 
    userId: user.id, 
    email: user.email,
    name: user.full_name 
  }, JWT_SECRET, { expiresIn: '24h' });
  
  console.log(`✅ Login successful for: ${user.full_name} (ID: ${user.id})`);
  
  res.json({
    success: true,
    message: 'Login successful',
    data: {
      user: {
        id: user.id,
        name: user.full_name,
        email: user.email,
        role: user.role
      },
      token: token
    }
  });
});

// ===== FIX 2: PROFILE API - Dynamic, not hardcoded Jenny Wilson =====
app.get('/api/me/profile', authenticateToken, (req, res) => {
  const userId = req.user?.userId || 1;
  const user = persistentUsers[userId];
  
  if (!user) {
    return res.status(404).json({
      success: false,
      message: 'User not found'
    });
  }
  
  console.log(`👤 Profile requested for: ${user.full_name} (ID: ${userId})`);
  
  res.json({
    success: true,
    message: "Profile information retrieved successfully",
    data: {
      user: {
        id: user.id,
        first_name: user.first_name,
        last_name: user.last_name,
        full_name: user.full_name,
        email: user.email,
        phone: user.phone,
        profile_photo: user.profile_photo,
        role: user.role,
        company: "ACME Inc.",
        joined_date: "August 17, 2025",
        employee_id: `EMP00${user.id}`,
        department: user.role === "Developer" ? "Engineering" : user.role === "Designer" ? "Design" : "Management",
        status: "Active",
        timezone: "UTC-5",
        last_login: new Date().toISOString()
      },
      permissions: {
        can_edit_profile: true,
        can_change_password: true,
        can_delete_account: true,
        can_upload_photo: true
      }
    }
  });
});

// ===== FIX 3: PROFILE UPDATE API - Persistent changes =====
app.put('/api/me/profile', authenticateToken, (req, res) => {
  const userId = req.user?.userId || 1;
  const { first_name, last_name, email, phone } = req.body;
  
  console.log(`📝 Profile update for user ${userId}:`, { first_name, last_name, email });
  
  if (!persistentUsers[userId]) {
    return res.status(404).json({
      success: false,
      message: 'User not found'
    });
  }
  
  // Update user data
  if (first_name) persistentUsers[userId].first_name = first_name;
  if (last_name) persistentUsers[userId].last_name = last_name;
  if (email) persistentUsers[userId].email = email;
  if (phone) persistentUsers[userId].phone = phone;
  
  // Update full name
  persistentUsers[userId].full_name = `${persistentUsers[userId].first_name} ${persistentUsers[userId].last_name}`;
  
  // Save to persistent storage
  savePersistentData();
  
  console.log(`✅ Profile updated successfully: ${persistentUsers[userId].full_name}`);
  
  res.json({
    success: true,
    message: "Profile updated successfully",
    data: {
      user: persistentUsers[userId],
      persistent: true
    }
  });
});

// ===== FIX 4: TIMER START API - Persistent, won't auto-stop =====
app.post('/api/me/timer/start', authenticateToken, (req, res) => {
  const userId = req.user?.userId || 1;
  const { projectId, locationId, notes } = req.body || {};
  const timerId = `timer_${userId}_${Date.now()}`;
  const startTime = new Date().toISOString();
  const today = new Date().toISOString().split('T')[0];
  
  console.log(`🚀 Timer start request - User: ${userId}, Timer: ${timerId}`);
  
  // Check if already has active timer
  const existingTimer = persistentTimers[userId];
  if (existingTimer && existingTimer.isActive) {
    return res.status(400).json({
      success: false,
      message: 'Timer already running. Stop current timer first.',
      data: {
        currentTimer: existingTimer,
        hasActiveTimer: true
      }
    });
  }
  
  // Check daily limit
  if (dailyLimits[userId] === today) {
    return res.status(400).json({
      success: false,
      message: 'Work session already completed for today. You can start again tomorrow.',
      data: {
        stoppedToday: true,
        canStartAgain: false
      }
    });
  }
  
  // Create persistent timer
  persistentTimers[userId] = {
    timerId,
    userId,
    startTime,
    projectId,
    locationId,
    notes,
    isActive: true,
    isRunning: true,
    isPaused: false,
    totalPausedTime: 0,
    pauseStartTime: null,
    createdAt: new Date().toISOString(),
    lastActivity: new Date().toISOString()
  };
  
  // Save immediately
  savePersistentData();
  
  console.log(`✅ Timer started successfully for user ${userId}`);
  
  res.json({
    success: true,
    message: 'Timer started successfully',
    data: {
      timerId,
      startTime,
      isRunning: true,
      isPaused: false,
      projectId: projectId || null,
      locationId: locationId || null,
      notes: notes || '',
      persistent: true
    }
  });
});

// ===== FIX 5: TIMER CURRENT API - Always shows correct state =====
app.get('/api/me/timer/current', authenticateToken, (req, res) => {
  const userId = req.user?.userId || 1;
  
  console.log(`🔍 Timer status check for user ${userId}`);
  
  // Reload from persistent storage
  loadPersistentData();
  
  const timerData = persistentTimers[userId];
  const today = new Date().toISOString().split('T')[0];
  const stoppedToday = dailyLimits[userId] === today;
  
  if (!timerData || !timerData.isActive) {
    console.log(`❌ No active timer for user ${userId}`);
    
    return res.json({
      success: true,
      data: {
        hasActiveTimer: false,
        timer: null,
        stoppedToday,
        canStart: !stoppedToday,
        message: stoppedToday 
          ? 'Work session completed for today. You can start again tomorrow.' 
          : 'No active timer - ready to start work session'
      }
    });
  }
  
  console.log(`✅ Active timer found for user ${userId}: ${timerData.timerId}`);
  
  // Calculate duration
  const now = new Date();
  const startTime = new Date(timerData.startTime);
  let currentTimeMs = now - startTime - (timerData.totalPausedTime || 0);
  
  if (timerData.isPaused && timerData.pauseStartTime) {
    const pauseDuration = now - new Date(timerData.pauseStartTime);
    currentTimeMs -= pauseDuration;
  }
  
  const currentHours = Math.max(0, currentTimeMs / (1000 * 60 * 60));
  const hours = Math.floor(currentHours);
  const minutes = Math.round((currentHours - hours) * 60);
  const currentDuration = `${hours}h ${minutes}m`;
  
  res.json({
    success: true,
    data: {
      hasActiveTimer: true,
      timer: {
        timer_id: timerData.timerId,
        start_time: timerData.startTime,
        current_duration: currentDuration,
        is_running: timerData.isRunning && !timerData.isPaused,
        is_paused: timerData.isPaused,
        project_id: timerData.projectId,
        notes: timerData.notes
      }
    }
  });
});

// ===== FIX 6: TIMER PAUSE API - Fully implemented =====
app.post('/api/me/timer/pause', authenticateToken, (req, res) => {
  const userId = req.user?.userId || 1;
  const { action } = req.body; // 'pause', 'resume', or undefined for toggle
  
  console.log(`⏸️ Timer pause/resume request for user ${userId} - action: ${action || 'toggle'}`);
  
  const timerData = persistentTimers[userId];
  if (!timerData || !timerData.isActive) {
    return res.status(400).json({
      success: false,
      message: 'No active timer to pause/resume'
    });
  }
  
  const now = new Date();
  const isCurrentlyPaused = timerData.isPaused || false;
  
  // Determine action
  const shouldPause = action === 'pause' ? true : action === 'resume' ? false : !isCurrentlyPaused;
  
  if (shouldPause && !isCurrentlyPaused) {
    // Pause timer
    timerData.isPaused = true;
    timerData.pauseStartTime = now.toISOString();
    console.log(`⏸️ Timer paused for user ${userId}`);
  } else if (!shouldPause && isCurrentlyPaused) {
    // Resume timer
    if (timerData.pauseStartTime) {
      const pauseDuration = now - new Date(timerData.pauseStartTime);
      timerData.totalPausedTime = (timerData.totalPausedTime || 0) + pauseDuration;
    }
    timerData.isPaused = false;
    timerData.pauseStartTime = null;
    console.log(`▶️ Timer resumed for user ${userId}`);
  }
  
  timerData.lastActivity = now.toISOString();
  
  // Save immediately
  savePersistentData();
  
  res.json({
    success: true,
    message: `Timer ${timerData.isPaused ? 'paused' : 'resumed'} successfully`,
    data: {
      timerId: timerData.timerId,
      isPaused: timerData.isPaused,
      pausedAt: timerData.pauseStartTime,
      totalPausedTime: timerData.totalPausedTime || 0,
      status: timerData.isPaused ? 'paused' : 'running'
    }
  });
});

// ===== FIX 7: TIMER STOP API - Proper cleanup =====
app.post('/api/me/timer/stop', authenticateToken, (req, res) => {
  const userId = req.user?.userId || 1;
  const today = new Date().toISOString().split('T')[0];
  
  console.log(`⏹️ Timer stop request for user ${userId}`);
  
  const timerData = persistentTimers[userId];
  if (!timerData || !timerData.isActive) {
    return res.status(400).json({
      success: false,
      message: 'No active timer to stop'
    });
  }
  
  // Calculate duration
  const now = new Date();
  const startTime = new Date(timerData.startTime);
  const durationMs = now - startTime - (timerData.totalPausedTime || 0);
  const hours = Math.floor(durationMs / (1000 * 60 * 60));
  const minutes = Math.round((durationMs % (1000 * 60 * 60)) / (1000 * 60));
  const duration = `${hours}h ${minutes}m`;
  
  // Mark as stopped today
  dailyLimits[userId] = today;
  
  // Clear timer
  delete persistentTimers[userId];
  
  // Save changes
  savePersistentData();
  
  console.log(`✅ Timer stopped for user ${userId} - Duration: ${duration}`);
  
  res.json({
    success: true,
    message: 'Timer stopped successfully',
    data: {
      timer_id: timerData.timerId,
      end_time: now.toISOString(),
      total_duration: duration,
      status: 'completed'
    }
  });
});

// Test endpoints for debugging
app.get('/api/test-users', (req, res) => {
  const users = Object.values(persistentUsers).map(user => ({
    id: user.id,
    name: user.full_name,
    email: user.email,
    role: user.role,
    token: jwt.sign({ userId: user.id, email: user.email }, JWT_SECRET, { expiresIn: '24h' })
  }));
  
  res.json({
    success: true,
    message: "Available test users with tokens",
    data: { users }
  });
});

app.get('/api/health', (req, res) => {
  res.json({
    success: true,
    message: 'CLIENT ISSUES FIXED - Server running perfectly!',
    status: 'All fixes implemented',
    fixes: [
      'Profile persistence - no more Jenny Wilson revert',
      'Timer persistence - no auto-stopping',
      'Pause API fully implemented',
      'Login/Profile email consistency',
      'Data persistence across restarts'
    ]
  });
});

const PORT = process.env.PORT || 3000;

app.listen(PORT, () => {
  console.log(`🚀 CLIENT FIX SERVER running on http://localhost:${PORT}`);
  console.log(`🔧 All client issues have been addressed:`);
  console.log(`   ✅ Profile data persistence (no more Jenny Wilson revert)`);
  console.log(`   ✅ Timer persistence (no auto-stopping)`);  
  console.log(`   ✅ Pause/Resume API fully implemented`);
  console.log(`   ✅ Login/Profile email consistency`);
  console.log(`   ✅ Data persistence across server restarts`);
  console.log(`📧 Test with: curl http://localhost:${PORT}/api/health`);
});

module.exports = app;