# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Complete Working Time Management API",
    "description": "Complete API collection with all 117 endpoints for Working Time Management System",
    "version": "3.0.0",
    "contact": {
      "email": "support@api-layer.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://api-layer.vercel.app",
      "description": "Production server"
    },
    {
      "url": "https://apilayer.vercel.app",
      "description": "Alternative Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/me/profile": {
      "get": {
        "summary": "Get My Profile",
        "description": "Get current user's complete profile",
        "tags": ["User Management"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "My profile retrieved successfully"
          }
        }
      },
      "put": {
        "summary": "Update My Profile",
        "description": "Update current user's profile information",
        "tags": ["User Management"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "email": {"type": "string", "format": "email"},
                  "phone": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "User Login",
        "description": "Authenticate user and get access token",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Login successful"},
          "401": {"description": "Invalid credentials"}
        }
      }
    },
...
    "/api/me/timer/start": {
      "post": {
        "summary": "Start Timer",
        "description": "Start work timer for user",
        "tags": ["Timer Management"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["projectId"],
                "properties": {
                  "projectId": {"type": "string"},
                  "taskId": {"type": "string"},
                  "description": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Timer started successfully"}
        }
      }
    },
    "/api/me/timer/stop": {
      "post": {
        "summary": "Stop Timer",
        "description": "Stop work timer for user",
        "tags": ["Timer Management"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Timer stopped successfully"}
        }
      }
    },
    "/api/me/timer/pause": {
      "post": {
        "summary": "Pause Timer",
        "description": "Pause work timer for user",
        "tags": ["Timer Management"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Timer paused successfully"}
        }
      }
    },
    "/api/me/timer/current": {
      "get": {
        "summary": "Get Current Timer",
        "description": "Get current running timer status",
        "tags": ["Timer Management"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Current timer status retrieved"}
        }
      }
    },
...
  }
}